added quick reply to rcs api
