+++ spec/components/schemas/content/nestable-contents/buttons-list.ts
