+++ spec/components/schemas/content/nestable-contents/share-location.ts
