+++ spec/components/schemas/content/nestable-contents/view-location.ts
