+++ spec/components/schemas/content/quick-reply.ts
@@ -1,7 +1,6 @@
 import { SchemaObject } from 'openapi3-ts';
 import { ref as baseRef } from './base';
 import { ref as quickReplyRef } from './nestable-contents/quick-reply-list';
-import { ref as buttonsRef } from './nestable-contents/buttons-list';
 import { createComponentRef } from '../../../../utils/ref';
 
 const buttons: SchemaObject = {
@@ -19,18 +18,6 @@ const buttons: SchemaObject = {
         type: 'string',
         description: 'Text to be sent',
         example: 'This is a text',
-      },
-      label: {
-        type: 'string',
-        description: 'Card title.',
-        example: 'example',
-      },
-      buttons: {
-        type: 'array',
-        description: 'Array of buttons displayed inside the card. The same used on Button Menu and Quick Reply.',
-        items: {
-          $ref: buttonsRef,
-        },
       },
       quickReplyButtons: {
         type: 'array',
