+++ spec/components/schemas/content/rcs/all.ts
@@ -3,6 +3,7 @@ import { ref as textRef } from '../text';
 import { ref as cardRef } from '../card';
 import { ref as fileRef } from '../file';
 import { ref as buttonRef } from '../buttons';
+import { ref as quickReplyRef } from '../quick-reply';
 import { createComponentRef } from '../../../../../utils/ref';
 
 const all: SchemaObject = {
@@ -13,6 +14,8 @@ const all: SchemaObject = {
   }, {
     $ref: buttonRef,
   }, {
+    $ref: quickReplyRef,
+  }, {
     $ref: cardRef,
   }],
   discriminator: {
@@ -21,6 +24,7 @@ const all: SchemaObject = {
       text: textRef,
       file: fileRef,
       button: buttonRef,
+      quick_reply: quickReplyRef,
       card: cardRef,
     },
   },
