+++ spec/tags/content-types.md
@@ -11,12 +11,12 @@ Messaging can be handled through SMS, WhatsApp, Facebook and RCS channels. For e
 | template     | &#10005; | &#10005; | &#10004; |          |
 
 ## Text
-This type of content is the most used type of content, and its composed of a plain text.
+This type of content is the most used one and is composed of plain text.
 
 <SchemaDefinition schemaRef="#/components/schemas/content.text" />
 
 ## File
-This is used to send a file to the user. The file will be presented to the user in different manner based on the file type. There are four types of presentation:
+This type of content is used to send a file to the user. Depending on the file type, the file itself will be displayed with a different appearance. There are four types of presentation:
 * Image
 * Video
 * Audio
@@ -25,17 +25,17 @@ This is used to send a file to the user. The file will be presented to the user 
 <SchemaDefinition schemaRef="#/components/schemas/content.file" />
 
 ## Contacts
-Use this type of content to send contacts information to the user.
+This type of content is used to send contact information to the user.
 
 <SchemaDefinition schemaRef="#/components/schemas/content.contacts" />
 
 ## Location
-This content is used to send location messages representing a point on the map to the user.
+This type of content is used to send location messages represented as a point on the map to the user.
 
 <SchemaDefinition schemaRef="#/components/schemas/content.location" />
 
 ## Template
-Template contents have a fixed text content, with a few variables placed where necessary. This type of content must be submitted for approval to WhatsApp. The nature of the content that is subjected to approval is very limited.
+This type of content has an underlying fixed text content with some required variables. After being filled, the template must be submitted for approval to WhatsApp. The rules that the submitted content must abide to are very restrictive.
 
 <SchemaDefinition schemaRef="#/components/schemas/content.template" />
